Write full test suite for new behaviour
